% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_computations.R
\name{check_alarm}
\alias{check_alarm}
\title{Compuates if an alarm needs to be generated based on the positive derivative of the slope}
\usage{
check_alarm(dat, ...)
}
\arguments{
\item{dat}{DataFrame containing x = time, y = RMS values}

\item{window_width}{The interval over which to calculate the slope}

\item{time_limit}{Limit over time (default = 10s)}

\item{threshold}{Cut off for change that is indicative of alarm (default = 0.01)}
}
\description{
Compuates if an alarm needs to be generated based on the positive derivative of the slope
}
\examples{
compute_alarm(df, 100)
}
