% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_computations.R
\name{compute_alarm}
\alias{compute_alarm}
\title{Compuates and plot if an alarm needs to be generated based on the positive derivative of the slope}
\usage{
compute_alarm(rms_data, plot_graph = T, ...)
}
\arguments{
\item{rms_data}{DataFrame containing x = time, y = RMS values}

\item{plot_graph}{T/F for plotting alarm graph}

\item{window_width}{The interval over which to calculate the slope}

\item{time_limit}{Limit over time (default = 10s)}

\item{x_lab}{x-axis label for graph, if plot_graph = T}

\item{y_lab}{y-axis label for graph, if plot_graph = T}

\item{title}{Title for graph, if plot_graph = T}
}
\description{
Compuates and plot if an alarm needs to be generated based on the positive derivative of the slope
}
\examples{
compute_and_plot_alarm(df, 100)
}
