% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_computations.R
\name{compute_feautres}
\alias{compute_feautres}
\title{Compuates ts features of a given window size using ACM package}
\usage{
compute_feautres(rms_data, window_size = 50, norm = T)
}
\arguments{
\item{rms_data}{DataFrame of RMS values}

\item{window_size}{Window size}

\item{norm}{= T, Normalise the features}
}
\description{
Compuates ts features of a given window size using ACM package
}
