% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_computations.R
\name{compute_second_derivative}
\alias{compute_second_derivative}
\title{Compuates second derivative of the time series}
\usage{
compute_second_derivative(dat, window_size)
}
\arguments{
\item{window_size}{Window size}

\item{x}{DataFrame of time series RMS value}
}
\description{
Compuates second derivative of the time series
}
