% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{convert_channel_data_csv}
\alias{convert_channel_data_csv}
\title{Saves the PNRF file channel data as a csv file}
\usage{
convert_channel_data_csv(file_path, save_path, channel_no, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{save_path}{Path to the converted csv file location, full path or current directory}

\item{channel_no}{Channel number to retrieve data}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}

\item{sampling}{If sampling is required (default = F)}

\item{sampling_rate}{Rate of sampling (default = 100), applicable if sampling = T}
}
\description{
Saves the PNRF file channel data as a csv file
}
\examples{
convert_channel_data_csv("example.PNRF","modified_example.csv", 3)
}
