% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_computations.R
\name{generate_model}
\alias{generate_model}
\title{Generates a SVM model}
\usage{
generate_model(dat, cross_validatation_type)
}
\arguments{
\item{dat}{DataFrame of time series dat}

\item{cross_validatation_type}{Cross_validatation_type # todo}
}
\description{
Generates a SVM model
}
