% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_all_channel_data}
\alias{read_all_channel_data}
\title{Read all channel data from PNRF file
Always reads sampled data}
\usage{
read_all_channel_data(file_path, sampling_rate = 100, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{sampling_rate}{Rate of sampling (default = 100)}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}
}
\value{
PNRFR file object
}
\description{
Read all channel data from PNRF file
Always reads sampled data
}
\examples{
read_all_channel_data("example.PNRF")
}
