% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_channel_data}
\alias{read_channel_data}
\title{PNRF file channel data}
\usage{
read_channel_data(file_path, channel_no, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to read}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}

\item{sampling}{If sampling is required (default = F)}

\item{sampling_rate}{Rate of sampling (default = 100), applicable if sampling = T}
}
\description{
PNRF file channel data
}
\examples{
read_channel_data("example.PNRF", 3)
read_channel_data("example.PNRF", 3, sampling = T, sampling_rate = 1000)
}
