% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_channel_data_obs}
\alias{read_channel_data_obs}
\title{PNRF file channel data size.}
\usage{
read_channel_data_obs(file_path, channel_no, data_type = 3)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to read}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}
}
\value{
Total number of observations in the channel specified (int)
}
\description{
Reads the total no of observations in a given channel.
}
\examples{
read_channel_data_obs("example.PNRF", channel_no=3)
read_channel_data_obs("example.PNRF", channel_no=3, data_type = 1)
}
