% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_channel_data_size}
\alias{read_channel_data_size}
\title{PNRF file channel data size.}
\usage{
read_channel_data_size(file_path, channel_no, data_type = 3)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to read}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}
}
\description{
Reads the PNRF file's specified channel data size.
}
\examples{
read_channel_data_size("example.PNRF", 3)
}
