% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_file_meta}
\alias{read_file_meta}
\title{PNRF file meta details.}
\usage{
read_file_meta(file_path, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory.}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}
}
\description{
Returns the file meta information: name, record details and channel details.
}
\examples{
read_file_meta("example.PNRF")
read_file_meta("E:\\\\temp\\\\example.PNRF")
}
