% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_segment_data}
\alias{read_segment_data}
\title{Reads the segment data corresponding to a channel}
\usage{
read_segment_data(file_path, channel_no, segment_no, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to retrieve data}

\item{segment_no}{Segment number of the corresponding channel}

\item{data_type}{Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}

\item{sampling}{If sampling is required (default = F)}

\item{sampling_rate}{Rate of sampling (default = 100), applicable if sampling = T}
}
\value{
PNRFR file object
}
\description{
Reads the segment data corresponding to a channel
}
\examples{
read_channel_data("example.PNRF", channel_no = 3, segment_no = 101)
read_channel_data("example.PNRF", channel_no = 3, segment_no = 101, sampling = T)
read_channel_data("example.PNRF", channel_no = 3, segment_no = 101,
sampling = T, sampling_rate = 1000)
}
