% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_computations.R
\name{compute_moving_average}
\alias{compute_moving_average}
\title{Computes the moving average of a given channel}
\usage{
compute_moving_average(file_path, channel_no, window_time, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to retrieve data}

\item{window_time}{Moving average time interval}

\item{...}{sampling_rate: Rate of sampling (default = 100), applicable if sampling = TRUE}
}
\value{
PNRFR file object
}
\description{
Computes the moving average of a given channel
}
