% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_channel_data}
\alias{read_channel_data}
\title{Read the channel data from a PNRF file}
\usage{
read_channel_data(file_path, channel_no, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to read}

\item{...}{sampling_rate: Rate of sampling (default = 100), applicable if sampling = TRUE}
}
\value{
PNRFR file object
}
\description{
Read the channel data from a PNRF file
}
\examples{
result <- read_channel_data(system.file("extdata", "VT094.pnrf", package="pnrfr"), 3)
result <- read_channel_data(system.file("extdata", "VT094.pnrf", package="pnrfr"), 3,
sampling = TRUE, sampling_rate = 1000)
}
