% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_file_meta}
\alias{read_file_meta}
\title{PNRF file meta details.}
\usage{
read_file_meta(file_path, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or file name in current directory.}

\item{...}{data_type: Type of data to retrieve : continuous 1, sweeps = 2, mixed = 3 (default = 3)}
}
\value{
A list containing file meta-information: title, recordCount, channelCount, records(list)
}
\description{
Returns the file meta information: name, record details and channel details.
}
\examples{
result <- read_file_meta(system.file("extdata", "VT094.pnrf", package="pnrfr"))
result <- read_file_meta(system.file("extdata", "VT094.pnrf", package="pnrfr"), data_type = 1)
}
