% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{read_segment_data}
\alias{read_segment_data}
\title{Reads the segment data corresponding to a channel}
\usage{
read_segment_data(file_path, channel_no, segment_no, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{channel_no}{Channel number to retrieve data}

\item{segment_no}{Segment number of the corresponding channel}

\item{...}{sampling_rate: Rate of sampling (default = 100), applicable if sampling = TRUE}
}
\value{
PNRFR file object
}
\description{
Reads the segment data corresponding to a channel
}
\examples{
result <- read_channel_data(system.file("extdata", "VT094.pnrf", package="pnrfr"),
channel_no = 3, segment_no = 101)
result <- read_channel_data(system.file("extdata", "VT094.pnrf", package="pnrfr"),
channel_no = 3, segment_no = 101, sampling = TRUE)
result <- read_channel_data(system.file("extdata", "VT094.pnrf", package="pnrfr"),
channel_no = 3, segment_no = 101, sampling = TRUE, sampling_rate = 1000)
}
