% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrf_reader.R
\name{save_channel_data.to_csv}
\alias{save_channel_data.to_csv}
\title{Saves channel data as a csv file}
\usage{
save_channel_data.to_csv(file_path, save_path, channel_no, ...)
}
\arguments{
\item{file_path}{Path to the PNRF file location, full path or current directory}

\item{save_path}{Path to the converted csv file location, full path or current directory}

\item{channel_no}{Channel number to retrieve data}

\item{...}{sampling_rate: Rate of sampling (default = 100), applicable if sampling = TRUE}
}
\value{
Noneo
}
\description{
Saves channel data as a csv file
}
\examples{
save_channel_data.to_csv(system.file("extdata", "VT094.pnrf", package="pnrfr"),"VT094.csv", 3,
sampling = TRUE, sampling_rate = 1000)
}
